[handler] randomize passwords in K:H:CreateDomain
With this patch K:H:CreateDomain gets the option pwrand where you can specify
a list of users to randomize their password. The randomized passwords are
stored in the database and this filed can be encrypted optionally by specifying
a recipients list.

This feature should avoid security flaws caused by default passwords on the
images we use (which need default passwords for certain users)